feat: add variables lib
